+++ inc/admin/auth.php
@@ -253,6 +253,7 @@ function wolf_core_wrong_theme() {
 		'morvan',
 		'clix',
 		'gainlab',
+		'unimate',
 	);
 
 	return ( ! in_array( esc_attr( sanitize_title_with_dashes( get_template() ) ), $ok, true ) );